requirements.txt
pyspark==3.5.0
mysql-connector-python>=8.0
pandas
pytest
requests
ipykernel
